// CleanFix-frontend/src/app/core/infrastructure/api/cleanfix-api/completedtask-repository/completedTaksDto.ts
export interface CompletedTaskDto {
  id: number
  address: string
  creationDate: Date
  completionDate: Date
  price: number
  issueType: string
  isSolicitation: boolean
  surface: number
  materials: string[]
  company: {
    name: string
    address: string
    number: string
    email: string
  }
}